[Refactor]: subcardsViewController내부에 있는 scrollView 분리
