+++ TripLog/SubCardsView/Delegate/MainCardDelegate.swift
@@ -8,5 +8,7 @@
 import Foundation
 
 protocol MainCardDelegate: AnyObject {
+    
     func changeSubCards(mainCardId: UUID, card: [SubCard])
+    
 }
