+++ TripLog/SubCardsView/SubCardEditViewController.swift
@@ -13,14 +13,7 @@ final class SubCardEditViewController: UIViewController {
     private let mainQueue = DispatchQueue.main
     private var selctedCardId: UUID?
     
-    private lazy var scrollView = UIScrollView()
-    private lazy var contentView = UIView()
-    private lazy var titleView = TitleView()
-    private lazy var imageView = SelectedImageView()
-    private lazy var starRateView = StarRateView()
-    private lazy var priceView = PriceView()
-    private lazy var locationView = LocationView()
-    private lazy var scriptTextView = UITextView()
+    private lazy var scrollView = SubCardScrollView()
     
     init(viewModel: SubCardsViewModel) {
         self.viewModel = viewModel
@@ -42,15 +35,7 @@ final class SubCardEditViewController: UIViewController {
         view.backgroundColor = .basic
         
         configureNavigationBar()
-        
         configureScrollView()
-        configureContentView()
-        configureTitleView()
-        configureImageView()
-        configureStarsView()
-        configurePriceView()
-        configureLocationView()
-        configureScriptView()
     }
 }
 
@@ -65,23 +50,7 @@ extension SubCardEditViewController {
         guard let index = index else { return }
         
         let card = viewModel.list.value[index]
-        
-        titleView.updateText(card.title)
-        
-        if let cardImages = card.images,
-            !cardImages.isEmpty {
-            imageView.images = cardImages
-            imageView.updateNumberOfPages(num: cardImages.count)
-            imageView.updateImage(image: cardImages[0])
-            imageView.isButtonHidden(value: true)
-        } else {
-            imageView.isButtonHidden(value: false)
-        }
-        
-        starRateView.starState = card.starsState
-        starRateView.updateButton()
-        priceView.updatePrice(price: card.money)
-        scriptTextView.text = card.script
+        scrollView.updateContent(card: card)
        
     }
 
@@ -96,32 +65,24 @@ extension SubCardEditViewController {
         if let cardId = selctedCardId {
             
             viewModel.updateContent(selectedCardId: cardId,
-                                    title: titleView.text,
-                                    images: imageView.images,
-                                    starsState: starRateView.starState,
-                                    price: priceView.price,
-                                    script: scriptTextView.text)
+                                    title: scrollView.titleView.text,
+                                    images: scrollView.imageView.images,
+                                    starsState: scrollView.starRateView.starState,
+                                    price: scrollView.priceView.price,
+                                    script: scrollView.scriptTextView.text)
             
         } else {
             
-            viewModel.appendSubCard(title: titleView.text,
-                                    images: imageView.images,
-                                    starsState: starRateView.starState,
-                                    price: priceView.price,
-                                    script: scriptTextView.text)
+            viewModel.appendSubCard(title: scrollView.titleView.text,
+                                    images: scrollView.imageView.images,
+                                    starsState: scrollView.starRateView.starState,
+                                    price: scrollView.priceView.price,
+                                    script: scrollView.scriptTextView.text)
         }
         
         navigationController?.popViewController(animated: true)
     }
-    
-    
-    @objc private func presentSearchLocationView() {
-    
-        let locationViewController = LocationViewController()
-        locationViewController.isModalInPresentation = false
-        self.present(locationViewController, animated: true)
-    }
-    
+
 }
 
 
@@ -142,7 +103,8 @@ extension SubCardEditViewController {
         view.addSubview(scrollView)
         
         scrollView.translatesAutoresizingMaskIntoConstraints = false
-        
+        scrollView.configureDelegate(delegate: self)
+
         let scrollViewConstraints = [
             scrollView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
             scrollView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
@@ -152,174 +114,12 @@ extension SubCardEditViewController {
         
         NSLayoutConstraint.activate(scrollViewConstraints)
     }
-    
-    private func configureContentView() {
-        scrollView.addSubview(contentView)
-        
-        contentView.translatesAutoresizingMaskIntoConstraints = false
-        
-        let contentViewConstraints = [
-            contentView.topAnchor.constraint(equalTo: scrollView.topAnchor),
-            contentView.bottomAnchor.constraint(equalTo: scrollView.bottomAnchor),
-            contentView.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor),
-            contentView.trailingAnchor.constraint(equalTo: scrollView.trailingAnchor),
-            contentView.widthAnchor.constraint(equalTo: scrollView.widthAnchor)
-        ]
-        
-        NSLayoutConstraint.activate(contentViewConstraints)
-    }
-    
-    private func configureTitleView() {
-        contentView.addSubview(titleView)
-        
-        titleView.translatesAutoresizingMaskIntoConstraints = false
-        
-        titleView.layer.cornerRadius = 8
-        titleView.layer.borderWidth = 1
-        titleView.backgroundColor = .viewBackground
-        
-        let viewConstraints = [
-            titleView.topAnchor.constraint(equalTo: contentView.topAnchor,
-                                           constant: 16),
-            titleView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor,
-                                               constant: 16),
-            titleView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor,
-                                                constant: -16),
-            titleView.heightAnchor.constraint(greaterThanOrEqualTo: view.heightAnchor,
-                                              multiplier: 0.1)
-        ]
-        
-        NSLayoutConstraint.activate(viewConstraints)
-    }
-    
-    private func configureImageView() {
-        contentView.addSubview(imageView)
-        
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        
-        imageView.layer.borderWidth = 1
-        imageView.layer.cornerRadius = 8
-        imageView.backgroundColor = .viewBackground
-        imageView.delegate = self
-        imageView.clipsToBounds = true
-        
-        let imageViewConstraints = [
-            imageView.topAnchor.constraint(equalTo: titleView.bottomAnchor,
-                                           constant: 16),
-            imageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor,
-                                               constant: 16),
-            imageView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor,
-                                                constant: -16),
-            imageView.heightAnchor.constraint(lessThanOrEqualTo: view.heightAnchor,
-                                              multiplier: 0.3),
-            imageView.heightAnchor.constraint(greaterThanOrEqualTo: view.heightAnchor,
-                                              multiplier: 0.25)
-        ]
-        
-        NSLayoutConstraint.activate(imageViewConstraints)
-    }
-    
-    private func configureStarsView() {
-        contentView.addSubview(starRateView)
-        
-        starRateView.translatesAutoresizingMaskIntoConstraints = false
-        starRateView.layer.cornerRadius = 8
-        starRateView.layer.borderWidth = 1
-        starRateView.backgroundColor = .viewBackground
-        
-        let viewConstraints = [
-            starRateView.topAnchor.constraint(equalTo: imageView.bottomAnchor,
-                                              constant: 16),
-            starRateView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor,
-                                                  constant: 16),
-            starRateView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor,
-                                                   constant: -16),
-            starRateView.heightAnchor.constraint(greaterThanOrEqualTo: view.heightAnchor,
-                                                 multiplier: 0.05)
-        ]
-        
-        NSLayoutConstraint.activate(viewConstraints)
-        
-    }
-    
-    private func configurePriceView() {
-        contentView.addSubview(priceView)
-        
-        priceView.translatesAutoresizingMaskIntoConstraints = false
-        
-        priceView.backgroundColor = .viewBackground
-        priceView.layer.cornerRadius = 8
-        priceView.layer.borderWidth = 1
-        
-        let viewConstraints = [
-            priceView.topAnchor.constraint(equalTo: starRateView.bottomAnchor,
-                                           constant: 16),
-            priceView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor,
-                                               constant: 16),
-            priceView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor,
-                                                constant: -16),
-            priceView.heightAnchor.constraint(greaterThanOrEqualTo: view.heightAnchor,
-                                              multiplier: 0.05)
-        ]
-        
-        NSLayoutConstraint.activate(viewConstraints)
-    }
-    
-    private func configureLocationView() {
-        contentView.addSubview(locationView)
-        
-        locationView.translatesAutoresizingMaskIntoConstraints = false
-        
-        locationView.layer.cornerRadius = 8
-        locationView.layer.borderWidth = 1
-        locationView.backgroundColor = .viewBackground
-        locationView.addGestureRecognizer(UITapGestureRecognizer(target: self,
-                                                                 action: #selector(presentSearchLocationView)))
-        
-        let viewConstraints = [
-            locationView.topAnchor.constraint(equalTo: priceView.bottomAnchor,
-                                              constant: 16),
-            locationView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor,
-                                                  constant: 16),
-            locationView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor,
-                                                   constant: -16),
-            locationView.heightAnchor.constraint(greaterThanOrEqualTo: view.heightAnchor,
-                                                 multiplier: 0.05)
-        ]
-        
-        NSLayoutConstraint.activate(viewConstraints)
-    }
-    
-    private func configureScriptView() {
-        contentView.addSubview(scriptTextView)
-        
-        scriptTextView.translatesAutoresizingMaskIntoConstraints = false
-        
-        scriptTextView.font = .preferredFont(forTextStyle: .body)
-        scriptTextView.isScrollEnabled = false
-        scriptTextView.layer.cornerRadius = 8
-        scriptTextView.layer.borderWidth = 1
-        scriptTextView.backgroundColor = .viewBackground
-        
-        let viewConstraints = [
-            scriptTextView.topAnchor.constraint(equalTo: locationView.bottomAnchor,
-                                                constant: 16),
-            scriptTextView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor,
-                                                    constant: 16),
-            scriptTextView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor,
-                                                     constant: -16),
-            scriptTextView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
-            scriptTextView.heightAnchor.constraint(greaterThanOrEqualTo: view.heightAnchor, multiplier: 0.5)
-        ]
-        
-        NSLayoutConstraint.activate(viewConstraints)
-    }
-    
+
 }
 
-//MARK: - SelectedImageViewDelegate
+//MARK: - PresentViewDelegate
 
-extension SubCardEditViewController: SelectedImageViewDelegate {
+extension SubCardEditViewController: PresentViewDelegate {
     
     func presentViewController(where viewController: UIViewController) {
         self.present(viewController, animated: true)
