--- TripLog/ViewComponent/Delegate/SelectedImageDelegate.swift
@@ -1,13 +1,0 @@
-//
-//  SelectedImageDelegate.swift
-//  TripLog
-//
-//  Created by 최승범 on 2024/03/14.
-//
-
-import UIKit
-
-protocol SelectedImageViewDelegate: AnyObject {
-    
-    func presentViewController(where viewController: UIViewController)
-}
